Move private tf methods into class
