--- src/uk/gov/defra/ffc/Queues.groovy
@@ -1,4 +1,0 @@
-package uk.gov.defra.ffc
-
-class Queues implements Serializable {
-}
